[UMPA-256] Add : 반주 서비스 상세 페이지에 샘플 탭 추가
