+++ Umpa/Presentation/Main/TeacherFinder/ServiceDetail/MrCreationService/MrCreationServiceDetailView.swift
@@ -8,11 +8,11 @@ import SwiftUI
 struct MrCreationServiceDetailView: View {
   @InjectedObject(\.appState) private var appState
 
-  #if DEBUG
-    @Injected(\.stubServiceDetailInteractor) private var serviceDetailInteractor
-  #else
-    @Injected(\.serviceDetailInteractor) private var serviceDetailInteractor
-  #endif
+#if DEBUG
+  @Injected(\.stubServiceDetailInteractor) private var serviceDetailInteractor
+#else
+  @Injected(\.serviceDetailInteractor) private var serviceDetailInteractor
+#endif
 
   let service: MusicCreationService
 
@@ -36,7 +36,25 @@ struct MrCreationServiceDetailView: View {
     ZStack(alignment: .bottom) {
       ScrollView {
         VStack(spacing: fs(0)) {
-          Header(tabSelection: $tabSelection, service: service)
+          VStack(spacing: fs(20)) {
+            Header(tabSelection: $tabSelection, service: service)
+
+            BottomLineSegmentedControl(
+              tabItems.map(\.name),
+              selection: $tabSelection,
+              appearance: .appearance(
+                buttonWidth: fs(70),
+                activeColor: UmpaColor.main,
+                bottomLineHeight: fs(2),
+                bottomLineOffset: fs(11),
+                activeFont: .pretendardSemiBold(size: fs(14)),
+                inactiveFont: .pretendardRegular(size: fs(14))
+              )
+            )
+            .padding(.horizontal, fs(26))
+            .innerStroke(.black.opacity(0.1), edges: .bottom, lineWidth: fs(1))
+          }
+
           segmentedControlContent
             .frame(maxWidth: .infinity, maxHeight: .infinity)
         }
@@ -66,10 +84,70 @@ struct MrCreationServiceDetailView: View {
     case .serviceOverview:
       ServiceOverviewTabContent(service: service)
     case .samplePreview:
-      SamplePreviewTabContent(sampleMusics: service.sampleMusics)
+      SampleMusicPreviewTabContent(sampleMusicList: service.sampleMusics)
     case .review:
       ReviewTabContent(service: service.eraseToAnyService())
     }
+  }
+}
+
+private struct Header: View {
+  @Binding var tabSelection: Int
+
+  let service: MusicCreationService
+
+  private let dotSize: CGFloat = fs(1.5)
+
+  var body: some View {
+    VStack(spacing: fs(20)) {
+      thumbnail
+      VStack(alignment: .leading, spacing: fs(6)) {
+        Text(service.title)
+          .font(.pretendardBold(size: fs(20)))
+
+        HStack(spacing: fs(4)) {
+          Text(service.author.name)
+            .font(.pretendardRegular(size: fs(12)))
+            .foregroundStyle(UmpaColor.mediumGray)
+          spacingDot
+          StarRating(service.rating)
+        }
+
+        UnitPriceView.V1(
+          model: .init(price: service.price, unitType: .song),
+          appearance: .fromDefault(priceColor: .black, priceFontSize: fs(17))
+        )
+        .padding(.vertical, fs(4))
+      }
+      .frame(maxWidth: .infinity, alignment: .leading)
+      .padding(.horizontal, fs(30))
+    }
+    .background(.white)
+  }
+
+  var thumbnail: some View {
+    AsyncImage(url: service.thumbnail) { image in
+      image
+        .resizable()
+        .aspectRatio(contentMode: .fill)
+    } placeholder: {
+      Color.gray
+    }
+    .frame(maxWidth: .infinity, height: ServiceDetailConstant.thumbnailHeight)
+  }
+
+  var spacingDot: some View {
+    Circle()
+      .frame(width: dotSize, height: dotSize)
+      .foregroundStyle(UmpaColor.mediumGray)
+  }
+}
+
+private struct ServiceOverviewTabContent: View {
+  let service: MusicCreationService
+
+  var body: some View {
+    Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
   }
 }
 
@@ -96,7 +174,7 @@ extension MrCreationServiceDetailView {
 }
 
 #if DEBUG
-  #Preview {
-    MrCreationServiceDetailView(service: .sample0)
-  }
+#Preview {
+  MrCreationServiceDetailView(service: .sample0)
+}
 #endif
