--- Umpa/Presentation/Main/TeacherFinder/ServiceDetail/ScoreCreationService/Components/ScoreCreationServiceOverviewTabContent.swift
@@ -1,20 +1,0 @@
-// Created for Umpa in 2025
-
-import Domain
-import SwiftUI
-
-extension ScoreCreationServiceDetailView {
-    struct ServiceOverviewTabContent: View {
-        let service: ScoreCreationService
-
-        var body: some View {
-            Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-        }
-    }
-}
-
-#if DEBUG
-#Preview {
-    ScoreCreationServiceDetailView.ServiceOverviewTabContent(service: .sample0)
-}
-#endif
