+++ Umpa/Presentation/Main/TeacherFinder/ServiceDetail/ScoreCreationService/ScoreCreationServiceDetailView.swift
@@ -8,11 +8,11 @@ import SwiftUI
 struct ScoreCreationServiceDetailView: View {
   @InjectedObject(\.appState) private var appState
 
-  #if DEBUG
-    @Injected(\.stubServiceDetailInteractor) private var serviceDetailInteractor
-  #else
-    @Injected(\.serviceDetailInteractor) private var serviceDetailInteractor
-  #endif
+#if DEBUG
+  @Injected(\.stubServiceDetailInteractor) private var serviceDetailInteractor
+#else
+  @Injected(\.serviceDetailInteractor) private var serviceDetailInteractor
+#endif
 
   let service: ScoreCreationService
 
@@ -36,7 +36,25 @@ struct ScoreCreationServiceDetailView: View {
     ZStack(alignment: .bottom) {
       ScrollView {
         VStack(spacing: fs(0)) {
-          Header(tabSelection: $tabSelection, service: service)
+          VStack(spacing: fs(20)) {
+            Header(tabSelection: $tabSelection, service: service)
+
+            BottomLineSegmentedControl(
+              tabItems.map(\.name),
+              selection: $tabSelection,
+              appearance: .appearance(
+                buttonWidth: fs(70),
+                activeColor: UmpaColor.main,
+                bottomLineHeight: fs(2),
+                bottomLineOffset: fs(11),
+                activeFont: .pretendardSemiBold(size: fs(14)),
+                inactiveFont: .pretendardRegular(size: fs(14))
+              )
+            )
+            .padding(.horizontal, fs(26))
+            .innerStroke(.black.opacity(0.1), edges: .bottom, lineWidth: fs(1))
+          }
+
           segmentedControlContent
             .frame(maxWidth: .infinity, maxHeight: .infinity)
         }
@@ -66,10 +84,116 @@ struct ScoreCreationServiceDetailView: View {
     case .serviceOverview:
       ServiceOverviewTabContent(service: service)
     case .samplePreview:
-      SamplePreviewTabContent(sampleSheets: service.sampleSheets)
+      SampleSheetPreviewTabContent(sampleSheetList: service.sampleSheets)
     case .review:
       ReviewTabContent(service: service.eraseToAnyService())
     }
+  }
+}
+
+private struct Header: View {
+  @Binding var tabSelection: Int
+
+  let service: ScoreCreationService
+
+  private let dotSize: CGFloat = fs(1.5)
+
+  var body: some View {
+    VStack(spacing: fs(20)) {
+      thumbnail
+
+      VStack(alignment: .leading, spacing: fs(6)) {
+        Text(service.title)
+          .font(.pretendardBold(size: fs(20)))
+
+        HStack(spacing: fs(4)) {
+          Text(service.author.name)
+            .font(.pretendardRegular(size: fs(12)))
+            .foregroundStyle(UmpaColor.mediumGray)
+          spacingDot
+          StarRating(service.rating)
+        }
+
+        HStack(spacing: fs(10)) {
+          ForEach(service.majors, id: \.self) { major in
+            Text(major.name)
+              .font(.pretendardMedium(size: fs(12)))
+              .padding(.horizontal, fs(16))
+              .padding(.vertical, fs(6))
+              .background(Color(hex: "E5EEFF"), in: Capsule())
+          }
+        }
+        .padding(.vertical, fs(8))
+
+        UnitPriceView.V1(
+          model: .init(price: service.basePrice, unitType: .sheet),
+          appearance: .fromDefault(priceColor: .black, priceFontSize: fs(17))
+        )
+        .padding(.vertical, fs(2))
+      }
+      .frame(maxWidth: .infinity, alignment: .leading)
+      .padding(.horizontal, fs(30))
+    }
+    .background(.white)
+  }
+
+  var thumbnail: some View {
+    AsyncImage(url: service.thumbnail) { image in
+      image
+        .resizable()
+        .aspectRatio(contentMode: .fill)
+    } placeholder: {
+      Color.gray
+    }
+    .frame(maxWidth: .infinity, height: ServiceDetailConstant.thumbnailHeight)
+  }
+
+  var spacingDot: some View {
+    Circle()
+      .frame(width: dotSize, height: dotSize)
+      .foregroundStyle(UmpaColor.mediumGray)
+  }
+}
+
+private struct ServiceOverviewTabContent: View {
+  let service: ScoreCreationService
+
+  var body: some View {
+    Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+  }
+}
+
+private struct SampleSheetPreviewTabContent: View {
+  let sampleSheetList: [SampleSheet]
+
+  private let sheetImageCornerRadius: CGFloat = fs(10)
+
+  var body: some View {
+    VStack(spacing: fs(12)) {
+      IndexingForEach(sampleSheetList) { _, sampleSheet in
+        AsyncImage(url: sampleSheet.url) { image in
+          image
+            .resizable()
+            .aspectRatio(contentMode: .fit)
+        } placeholder: {
+          Color.gray.opacity(0.5)
+        }
+        .frame(maxWidth: .infinity, idealHeight: fs(380))
+        .fixedSize(horizontal: false, vertical: true)
+        .overlay {
+          // TODO: 워터 마크 추가
+          Color.blue.opacity(0.1)
+        }
+        .clipShape(RoundedRectangle(cornerRadius: sheetImageCornerRadius))
+        .innerRoundedStroke(
+          UmpaColor.lightLightGray,
+          cornerRadius: sheetImageCornerRadius,
+          lineWidth: fs(1)
+        )
+      }
+    }
+    .padding(.horizontal, fs(30))
+    .padding(.vertical, fs(22))
   }
 }
 
@@ -96,7 +220,7 @@ extension ScoreCreationServiceDetailView {
 }
 
 #if DEBUG
-  #Preview {
-    ScoreCreationServiceDetailView(service: .sample0)
-  }
+#Preview {
+  ScoreCreationServiceDetailView(service: .sample0)
+}
 #endif
